fix(provider): use uint64 for the timeout invariant assertions
The compile-time ordering assertions converted Duration margins to uint, whose
width is architecture-dependent. The margins are nanosecond counts, so a valid
120-second margin is 120,000,000,000 — well beyond a 32-bit uint. The
assertions therefore failed to compile for the 386 and arm targets in
.goreleaser.yml:

  internal/provider/timeouts.go:87:11: constant 120000000000 overflows uint
  internal/provider/client.go:110:16: constant 210000000000 overflows uint

Since goreleaser only runs on v* tags, this would have surfaced while cutting
a release rather than on the pull request.

uint64 holds any time.Duration while still rejecting negative constants, so the
invariant is enforced identically on 32- and 64-bit targets. Verified by
building every goos/goarch pair in the release matrix, and by confirming an
inverted default still fails to compile on both 386 and arm64.

+++ internal/provider/client.go
@@ -106,8 +106,9 @@ const (
 
 // The Retry-After cap is only meaningful if a capped wait still leaves room for
 // the retry itself. Enforced at compile time; see timeouts.go for the same
-// technique applied to the timeout defaults.
-const _ = uint(DefaultReadTimeout - maxRetryAfter - minRetryHeadroom)
+// technique applied to the timeout defaults, and for why this is uint64 rather
+// than uint.
+const _ = uint64(DefaultReadTimeout - maxRetryAfter - minRetryHeadroom)
 
 // newRetryBackOff builds the default exponential retry policy.
 //
